# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "School Management System",
    "summary": "Multi-tenant school management platform",
    "description": "EduCloud API -- identity, tenancy, RBAC, plans and invitations.\n\n**Tenant isolation** is enforced by PostgreSQL Row-Level Security. The `org` claim in your access token scopes every query at the database level; cross-tenant reads return 404, never 403, so resource existence is never disclosed.\n\n**School scoping** is a second, softer boundary inside the organization: a school-scoped membership sees only its own campus, while an organization owner spans all of them.\n\n**Permissions** are `resource:action` strings resolved per request from the caller's role. They are deliberately not embedded in the token, so revoking access takes effect on the very next request.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/platform/auth/login": {
      "post": {
        "tags": [
          "Platform"
        ],
        "summary": "Platform Login",
        "description": "Sign in as a platform operator (spec \u00a74.3A).\n\nTHERE IS NO REGISTRATION ROUTE, and there never will be. Accounts are minted by\n`python -m app.cli seed`, and credentials rotate through the CLI. An emailed\npassword reset for this role would reduce the platform's security to the security\nof one inbox -- and this role can read every school's records.",
        "operationId": "platform_login_api_v1_platform_auth_login_post",
        "parameters": [
          {
            "name": "x-token-transport",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Token-Transport"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlatformLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformAdminRead"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
    "/api/v1/platform/auth/logout": {
      "post": {
        "tags": [
          "Platform"
        ],
        "summary": "Platform Logout",
        "description": "Revoke the operator session before clearing platform cookies.",
        "operationId": "platform_logout_api_v1_platform_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
    "/api/v1/platform/auth/me": {
      "get": {
        "tags": [
          "Platform"
        ],
        "summary": "Platform Me",
        "operationId": "platform_me_api_v1_platform_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformAdminRead"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
...
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register",
        "description": "Create an account and its organization (spec \u00a74.3B).\n\nReturns 201 with NO tokens: login is blocked until the emailed link is followed.",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
    "/api/v1/auth/verify-email": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Verify Email",
        "description": "Consume a verification link and activate the account.",
        "operationId": "verify_email_api_v1_auth_verify_email_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
    "/api/v1/auth/resend-verification": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Resend Verification",
        "description": "Re-send the verification link.\n\nAlways the same response, whether or not the address exists or is already\nverified -- otherwise this becomes a cheap way to test which addresses are\nregistered.",
        "operationId": "resend_verification_api_v1_auth_resend_verification_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendVerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login",
        "description": "Authenticate and, if the context is unambiguous, issue tokens (spec \u00a74.3D).\n\nA user with several memberships gets `select_required: true` and NO access\ncookie; they must call `POST /auth/context` to choose. See `LoginResponse`.",
        "operationId": "login_api_v1_auth_login_post",
        "parameters": [
          {
            "name": "x-token-transport",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Token-Transport"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
    "/api/v1/auth/context": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Switch Context",
        "description": "Switch the active membership and re-issue tokens (spec \u00a74.3E).\n\nDeliberately reachable with a token that has NO active membership -- that is the\nstate a multi-membership user is in immediately after login, and it is the only\nendpoint that can move them out of it.",
        "operationId": "switch_context_api_v1_auth_context_post",
        "parameters": [
          {
            "name": "x-token-transport",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Token-Transport"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContextSwitchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipSummary"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
...
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Logout",
        "description": "Revoke this session and clear the cookies.",
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PlatformLoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Password"
          },
          "totp_code": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\d{6}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Totp Code"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "PlatformLoginRequest"
      },
...
      "VerifyEmailRequest": {
        "properties": {
          "token": {
            "type": "string",
            "maxLength": 512,
            "minLength": 16,
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "VerifyEmailRequest"
      }
    }
  }
}